faet: reqTxMonitor 구현
refactor: DB 칼럼 타입 및 필드명 변경

+++ src/main/java/com/baekho/bridgenet/domain/chain/dto/request/ChainAddRequestDTO.java
@@ -47,5 +47,5 @@ public class ChainAddRequestDTO {
 
     @NotNull
     @Min(1)
-    private BigInteger requiredConfirmations;
+    private Integer requiredConfirmations;
 }
