+++ src/main/java/com/baekho/bridgenet/domain/chain/dto/response/AdminChainDetailDTO.java
@@ -18,5 +18,5 @@ public class AdminChainDetailDTO {
     private BigInteger maxFeePerGas;
     private BigInteger maxPriorityFeePerGas;
     private BigInteger gasLimit;
-    private BigInteger requiredConfirmations;
+    private Integer requiredConfirmations;
 }
