+++ src/main/java/com/baekho/bridgenet/domain/chain/dto/response/ChainAddResponseDTO.java
@@ -18,5 +18,5 @@ public class ChainAddResponseDTO {
     private BigInteger maxFeePerGas;
     private BigInteger maxPriorityFeePerGas;
     private BigInteger gasLimit;
-    private BigInteger requiredConfirmations;
+    private Integer requiredConfirmations;
 }
