+++ src/main/java/com/baekho/bridgenet/domain/chain/entity/Chain.java
@@ -65,7 +65,7 @@ public class Chain {
     private BigInteger gasLimit;
 
     @Column(name = "required_confirmations", nullable = false)
-    private BigInteger requiredConfirmations;
+    private Integer requiredConfirmations;
 
     @OneToMany(
             mappedBy = "chain",
